Refactor by moving SendButton
