+++ presentation/src/main/java/com/pocs/presentation/view/common/button/SendButton.kt
@@ -1,0 +1,31 @@
+package com.pocs.presentation.view.common.button
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Send
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
+import com.pocs.presentation.R
+
+@Composable
+fun SendButton(
+    onClick: () -> Unit,
+    enabled: Boolean = true
+) {
+    IconButton(
+        enabled = enabled,
+        onClick = onClick
+    ) {
+        Icon(
+            imageVector = Icons.Filled.Send,
+            contentDescription = stringResource(R.string.save),
+            tint = if (enabled) {
+                MaterialTheme.colorScheme.primary
+            } else {
+                MaterialTheme.colorScheme.onBackground.copy(alpha = 0.4f)
+            }
+        )
+    }
+}
